clusterdeployment: support cluster install ref
ClusterDeployment now supports 2 ways for installing clusters,
- Managed ClusterProvision, and
- Reference to an k8s object that implements ClusterInstall contract.

When reference to clusterinstall implementation is used,

the controller uses the .spec.imageSetRef from the clusterinstall obj to
get the ClusterImageSet required to get information of release image
etc.

the controller then watches the clusterinstall reference to sync
information to the clusterdeployment object like,

All the conditions in clusterinstall contract are synced to
clusterdeployment with `ClusterInstall` prefix for type.

The `ClusterInstallFailed` condition on the clusterdeployment is copied
to `ProvisionFailed` condition.

The `ClusterInstallStopped` condition is copied to `ProvisionStopped`
condition, except when `ClusterInstallCompleted` condition is in **bad**
state where `ProvisionStopped` condition is set to true with reason
InstallAttemptsReached.

ClusterMetadata from clusterinstall implementations is copied to the
clusterdeployment only when all the information in the metadata is set
because clusterdeployment object only accepts the entire thing or
nothing.

When ClusterInstallCompleted condition becomes true, the controller set
the Installed on clusterdeployment to true.

When clusterdeployment Installed is set true, no more syncs from
clusterInstallRef occur.